Add XML Schema to validate config for the Java encoder/decoder generator
According to this schema, configuration data for the generator
contains a single element, <root>, with a single attribute,
"element", which is required.

For example, in the tutorial:

  <config:root element="quizz"/>

This indicates the top level element of the XML structure to encode.

This configuration is included into XML Schema files provided as input
to the tool, as an extension of the XML Schema. It is located at the
following level of the schema:

  /xsd:schema/xsd:annotation/xsd:appinfo

with the prefix "config" bound to the namespace
"http://eric.brechemier.name/2004/generation/encoder-decoder/config",
usually in the root <xsd:schema> element.

This configuration is optional, but a warning is printed when it is
omitted, e.g.

  Missing root element config, first element quizz is considered root.

if the first element defined in the schema is "quizz".

In this case, as indicated in the message, the first element defined
is considered to be the root element of the documents.