Add XML Schema which defines a small subset of XForms
I developed this schema during my internship at Expway,
while working on a mobile application, a browser for small
documents written in a custom profile of SVG and XForms.
It was used to create applications to display service guides
like a list of VOD movies or a TV programme guide.

The main part of the documents was in a custom subset of SVG Tiny,
that we called SVG Nano, "nano" being even smaller than tiny.

Each SVG document defined a template for a different screen of the
application. They were part of the app while the data to display
was downloaded separately.

XForms elements were used to fill these SVG templates with data.
We called this small profile XForms Nano.

A few extra XForms elements were required to complete the definition
of the data binding, and were defined in a separate schema,
"NanoXFormsExpWayExtension.xsd", with its own namespace.

The comments in this file show traces of the template from which
it was created. Notably, the comment at the top, which still
includes a blank author name to be completed:

  *** Credits ***
  Schema: Your Name, Month 2004

The comments below also provide hint of the process used to create
the schema:

1. a template schema was copy/pasted
2. the target namespace had to be customized in multiple places
   at the top of the file
3. then by following instructions provided in comments in the template,
   only the list of elements could be provided at first
4. by running an ant task, the definition of corresponding types were
   added automatically, where missing, with different kinds of structures
   generated, based on hints, for elements with and without attributes
5. the type definitions were further customized to complete the schema

I later added a separate ant task taking care of the first two steps,
generating the empty schema with the target namespace properly set. This
process was less prone to errors than copy/pasting a previous schema.