Add Same but Different utility based on Drix 2002
This utility is based on the design pattern called "Copie non conforme"
(Same but Different) described on page 43 of the book "XSLT Fondamental"
by Philippe Drix (2002).

This is a very common pattern, that I revisited to handle different cases.

I used this pattern in many different projects before and since,
with variations and evolutions from project to project.