Add Quick Start guide, in source and PDF formats
This Quick Start guide describes the process to get started with
the tool, from the download of the project to the generation of
the custom encoder and decoder for a custom data format.

It makes references to two separate documents:

  * "see Quick Ref", in step 2.2
  * "see Architecture", in step 3.2

I assume that "Architecture" refers to the UML schema in the slide
"Tool Architecture", page 11/14 of the document overview/overview.pdf.

As for "Quick Ref", it does not correspond to a formal document that
I could identity. It may have referred to an informal text document or
to a template schema showing by example all the supported constructs,
like the one now generated by the ant task createEmptySchema. It may also
correspond to an early version of the HTML page which became the third
chapter of the tutorial, "Chapter 3: XML Schema". This page, now located
at the path doc/tutorial/chapter3/index.html, describes in details the
supported XML Schema types and elements.

It is dated from October 28th, 2004. The oldest version I found
on disk was dated from a few days later, on November 3rd, 2004.
This is the date that I set in this commit while trying to restore
parts of the history of the project in this new git repository.

The project was originally versioned in a centralized CVS repository,
mainly for backup purpose. It was also distributed to parts of the
developers of the team by email, in ZIP archives.

I have long lost access to the CVS repository, and a local CVS checkout
only contains the latest version of the files, unlike in git. But the
last modification dates of the files were preserved from disk to disk
in my backup copies, as well as in the ZIP archives.

I have thus decided to add the files in this new git repository at their
last modification date, which gives a useful indication, for separate files
of their contents when last modified at a specific date in the past.

Checking out the repository at a particular commit, however, will not
reflect the state of the repository at that time: only the files last
modified before the date of the commit will be present, which does not
reflect the actual files present in the repository at that time.