Add XSLT transformation to indent XML
I have written this utility for a past project, back in 2003.
That initial version was in French. I translated it to English
on October 20th, 2004.

I have reused this utility in multiple projects as indenting XML
is often necessary to analyze its structure and the patterns within.
This comes in handy when preparing a transformation on input data
received from a third party, e.g. exported from a database.