Add XSLT transform which encodes XML into binary using Java extensions
The input is an XML document with the same structure as the binary file
expected in output. Since XSLT does not support binary encoding in output,
the main output of the transformation is text, but it is currently empty.
This text output may be used to provide descriptive information in the
future. The binary output is produced as a side effect of the transformation,
by calling extension functions implemented in Java in our class JavaEncoder.
This is a mechanism supported by the Xalan XSLT engine used to run this
XSLT transform.