Add tutorial with step by step examples of use
The tutorial describes the creation of a schema for the data of a simple
quiz game, with questions and answers. It is divided into 5 chapters,
which guide the reader through all the steps of the process to generate
an encoder and a decoder for the game data:

  * Chapter 1: description of the game and initial software setup
  * Chapter 2: customization of user properties and creation of empty schema
  * Chapter 3: definition of data types in the schema, assisted with tooling
  * Chapter 4: XML data encoding into binary, and test decoding
  * Chapter 5: generation of decoder calling semantic actions added in schema

The tutorial invites the reader to follow the instructions to create their
own schema and semantic actions. But it also includes example files which
show what is the expected result of each step.

The tutorial is written in HTML, with a separate page per chapter.