Add utility to reformat Java source code: Artistic Style version 1.15.3
This is a Windows executable, which runs from the executable astyle.exe.
The documentation file astyle.html describes the different options.

This Windows executable is used in the implementation of the task
'indentJavaUsingArtisticStyle' in the Ant build.xml file. The task
is used to indent the source code of the generated Java decoders.
It can be disabled with limited loss of functionality. An alternative
should be found to improve cross-platform support, however.

Project Home Page:
https://sourceforge.net/projects/astyle/

Downloaded from:
https://sourceforge.net/projects/astyle/files/astyle/
(version 1.15.3 is no longer available)

License:
GNU GPL (at the time), MIT (now)