Add XSLT transform from Iso-Java XML to Java source code
The input of this transformation is XML with the same structure as
Java code: elements for classes, blocks, loops, conditionals, etc.
This XML format similar to Java is called "Iso-Java" in this project.

Generated source code is indented according to the nesting level of
the Java instructions, with a fixed indent of 3 spaces at each level.

Note that this .xsl file contains the implementation of the transformation,
not just the header like in other transforms found in the same folder.