Add XSLT transform to generate test decoder from the XML Schema
The output of this transform is still XML, but with the same structure
as the Java code to generate. The input file is an XML Schema expected
to contain annotations for semantic actions to include in the decoder.
These annotations are not used in the test decoder, however, they are
completely ignored by this transform.

This is the header for the transformation. I have separated the body of
the implementation into a separate file, with the same name but a different
extension, .body.xsl.