Add XSLT transform which generates decoders in Java
Generated  decoders decode binary data and make decoded data available
to Java code, semantic actions, provided by the schema designer in
annotations attached to the definitions of the corresponding data items.

Empty annotations are first generated by an Ant task, 'addSemanticActions',
then filled in with code which uses the data in the context of the target
application. This process is described in "Chapter 5: Semantic Actions"
in the tutorial.

There are two different types of annotations:
  1. context annotations which define variable types and class imports
  2. event annotations which run code at different steps of the decoding

The annotations generated by 'addSemanticActions' contain comments with
examples of use in each case, which help to understand what kind of code
typically goes into each type of annotation.

The mental model for writing annotations is illustrated on slide 6/14
of the presentation doc/overview.pdf, "Inside the Factory". The idea is
that the decoding process is like a factory which takes binary data as
input and builds objects in the target application, e.g. monsters, levels
and maps in a game. These objects are built in parts of different sizes,
which can take advantage of the hierarchical structure of the input data.
Parts built for child elements are made available in the code that assembles
the parent element. Similarly, parts constructed by the parent are provided
for enrichment and decoration in the code for each child element. This
concept is called "continuation" from child and from parent in the
annotations and the illustration of the mental model.

The decoding process follows a push model: data is processed by an external
parser which makes it available by pushing it too the user code at different
steps (or events) of the decoding. The push model is typically less
intuitive to use than a pull model, where the user code has all its context
in a single place from which it pulls the parts of the data that it needs
from a decoder used as a library. However, the system of continuation from
child and from parent mitigates this issue and makes this process more
usable than a typical push decoding process.

This transform extends the base decoder generator, which is shared
by the XSLT transforms which generate decoders and test decoders.

This is the header part of the transform. The implementation is found
in the associated .body.xsl file.