Add XSLT transform, common base for decoders and test decoders
All the common parts, browsing and generation of code structure,
as well as all common Java instructions found in both the decoder
and the test decoder are handled in this base transform.

The base transform is imported in the more specific transforms,
which override a list of base templates to generate the parts
specific to the decoder and the test decoder respectively.

These base templates to override are defined as empty templates
at the end of the base transform. Running the base transform thus
results in empty sections in all these specific parts, which get
filled with expected Java code in the decoders and the test decoders.

This is the header part of the transform. The implementation is
found in the associated .body.xsl file.