Add XSLT transform from XSL Schema to a generated XSLT encoder
This transform relies on the base transform XsdBrowsingTemplateMethod
to browse the input XML Schema. The XML Schema browsing part is common
to the transforms which generate custom encoders and decoders for XML
data that conforms to the schema.

In the specific steps of this transform, XSLT instructions are generated,
which will in turn transform XML data that conforms to the schema into
the "Iso-B4j" format which has the same structure as the binary file
expected at the end of the encoding chain.

It is the job of the Simple Encoder, B4jSimpleEncoder.xsl, to convert
the Iso-B4j XML into binary, through extension functions in Java loaded
and executed by the XSLT engine.

The different steps in this chain of transformations are shown on the
slide 12/14 of the presentation doc/overview/overview.pdf.

This .xsl file contains only the header of the transform;
the implementation is found in the associated .body.xsl file.