Add XML Schema with the definition of custom elements for Semantic Actions
The Semantic Actions are initially inserted into the user's schema
with the Ant task 'addSemanticActions'. This process is described
in chapter 4 of the tutorial.

An annotation is inserted in the root element of the schema, and
another in each of the complex types defined within. The comments
added to these generated annotions clarify their indented use.

Most of these annotations contain Java source code which will be
inserted at strategic locations in the decoder class generated
from the user's schema.

In the top-level annotation:

  - a4j:packages contains the Java code which defines packages to import
  - a4j:globalDef defines global, static attributes and methods for the decoder
  - a4j:events
    - a4j:context
      - a4j:fromParent: optional, Java code to declare variables which will
        be provided to the top-level decoding function
      - a4j:parentContination: optional, type and name of a single return
        value to be returned from the top-level decoding function
      - a4j:childContinuation: optional, list of parameters that will be
        provided in the call to the decoding function for each child element
    - a4j:onDocumentStart: Java code to run at the very start of the decoding
    - a4j:onEachChildEnd: optional, Java code to run after decoding each child
    - a4j:onDocumentEnd: Java code to run at the very end of the decoding
    - a4j:onError: optional, Java code to call when decoding fails

In each Complex Type annotations:

  - a4j:events
    - a4j:context
      - a4j:fromParent: optional, Java code to declare variables which will
        be provided by the parent element decoding function
      - a4j:parentContination: optional, type and name of a single return
        value which will be returned to the parent element decoding function
      - a4j:childContinuation: optional, list of parameters that will be
        provided in the call to the decoding function for each child element
    - a4j:onElementStart: Java code to run before decoding the element
    - a4j:onAttributesEnd: Java code to run after decoding all the attributes
    - a4j:onValueEnd: Java code to run after decoding the value of the element
    - a4j:onChildrenCount: optional, Java code to run after counting child
      elements, typically to initialize an array of the correct size
    - a4j:onEachChildEnd: optional, Java code to run after decoding each child
    - a4j:onElementEnd: Java code to run after decoding the element
  - a4j:topic: optional, semantic action for Id/Ref matches in given topic
    - a4j:context
      - a4j:fromReferree: declaration of type and name for published objects
      - a4j:fromReferrer: declaration of type and name for subscriber object
    - a4j:onMessage: Java code to run when the match occurs (see details below)

The a4j:context annotation may also contain a list of:
      - a4j:publish: name of a topic and an object related to an Id value,
        called the "referree"
      - a4j:subscribe: name of a topic and an object related to an IdRef value,
        called the "referrer"

Each time an IdRef value matches an Id value for the same topic,
the semantic action associated with the topic fires with both values.
The use case is to allow an object to link to shared parts using the id/ref
mechanism, as an extension to building from children parts own only by itself.
Typically, when the match occurs, the "referrer" object will store a reference
to the object created for the associated id, the "referree".

This mechanism is illustrated on slide 7/14 of the overview.pdf presentation.
This is an advanced feature, not part of the core B4J tool chain, but defined
as an extension called the powerpack.