Add XSLT Transform to indent file and remove a special temporary attribute
In the inline XSLT Transform src/schema/emptySchema.xsd (not just
an XML Schema file), a special attribute is created to force the
resulting file to include a declaration for the custom namespace
chosen by the user and the associated prefix, defined in two user
properties loaded by Ant build: namespace.prefix and namespace.uri.

The empty schema template adds a temporary attribute with the given
prefix, associated with the namespace and the text value "ignore".

This transform runs just after the creation of the empty schema,
in the Ant build task 'createEmptySchema', to remove the temporary
attribute while preserving the custom namespace declaration created
as a side effect.