Copy the XML Schema which defines Id/IdRef/includeFile in 'ignore' mode
This version of the codecs for Id/IdRef do not include the textual values
of the identifiers, only numeric values used for matching during decoding.
Both idRefInclude.xsd and idRefIgnore.xsd are imported in NanoSvg.xsd, but
only idRefIgnore.xsd is used. It has in effect replaced idRefInclude.xsd
which was used previously since the string values of the identifiers have
little use beyond the matching when decoding the binary file on a device.