Add XML Schema to define Id/IdRef without including them in binary output
This XML Schema is similar to idRefInclude.xsl found in the same folder:
it defines simple types for Id and IdRef in the same way. The difference
is only the namespace, which is processed differently in the encoder
generator. When the Id/IdRef come from the 'include' version of the
namespace, the values of the Id and IdRef are included in the output,
together with shorter numeric identifiers used for Id/IdRef matching
during decoding. With the 'ignore' version, only the numeric values are
included in the binary file, while the string values of the identifiers
are not.

This schema also defines a third simple type, ignore:includeFile,
for lack of a better place at that moment. The idea is also to use
input data, the path to a file name, to include something into the
binary file, the contents of that file, without including the string
value of the path itself. This definition should be moved to its own
file, if not in a different namespace.

This code has been implemented in the powerpack extension, in the Java
class name.brechemier.eric.PowerPackJavaEncoder defined in the folder
src/powerpack/encoder/java/name/brechemier/eric/PowerPackJavaEncoder.java.