Add Java class with codec for includeFile simple type
This codec is part of the "powerpack" extension. The includeFile simple
type is defined in src/schema/idRefIgnore.xsd. It allows to inline a
whole file, given its path location. Only the contents of the file is
included, not the path location.

This custom class defines the static method writeFile(String fileName)
which is called as an extension function in the XSLT transform from
Iso-B4J to binary part of the powerpack:

  src/powerpack/encoder/B4jSimpleEncoder.xsl