Add XML Schema for a subset of SVG Tiny nicknamed 'Nano SVG'
This is not an official SVG profile, but a limited subset with extensions
sufficient to develop the user interface of simple apps running in a custom
SVG browser on a mobile phone.

This version of the schema defines the supported grammar for these
SVG documents, without the semantic actions, which only contain
template comments generated by the Ant task 'addSemanticActions'
at this point.