Add implementation of the XSLT Transform which generates decoders
This transforms takes an XML Schema with Semantic Actions in
annotations as input and generates instructions for a Java decoder,
written in an XML format similar to the structure of the Java program,
as an intermediate step.

This implementation is relatively simple, due to the fact that all the
common parts for the generation of a decoder have been abstracted away
in the base XSLT Transform XsdToJavaDecoderBase.xsl, which is imported
in the header of this transform.

The responsibility of this transform is limited to copying Java code
from the semantic actions corresponding to the different kinds of
decoding events.