Add XSLT Transform which extends the XML Schema browser for the powerpack
Like the base XML Schema browser, this transform takes an XML Schema with
annotations for semantic actions as input and browses the structure of the
file, delegating further processing to more specific transforms.

This extension of the browsing transform, part of the powerpack extension,
handles specific features of the powerpack: Id/IdRef resolution, inline
files and extended codecs for simple types.