Add XSLT Transform which generates XSLT encoders for the powerpack
The implementation of extended codecs for the powerpack is done
completely in the base transform:

  src/powerpack/browsing/XsdBrowsingTemplateMethod.xsl

which handles Id/IdRef values to include or omit in the binary files.

The implementation of inline files done in this transform is straight-
forward as well, because the actual inlining is left to the next step,
which converts Iso-B4J files in XML format to binary.

Most of this transform is responsible for handling the Id/IdRef
resolution, with the use of two global stacks for the resolution of
all the Id and IdRef, in any topic, in the document:

  * backwardIds: stores publisher objects when Id comes before IdRefs
  * forwardRefs: stores subscriber objects when Id comes after IdRefs

In both cases, the role of these stacks is store a reference to the
object currently under construction, to provide it at a later point
when it matches another element built later in the document order.

These stacks are operated through abstract instructions defined in a
separate namespace associated with the prefix 'stack':

  * stack:new: create an empty stack
  * stack:putOn: add a value on the stack
  * stack:getPositionOn: look up position of a single value on the stack
    (used to locate the position of the Id matching an IdRef)
  * stack:getEachPositionOn: look up all positions of a value on the stack
    (used to locate the positions of all IdRef values matching an Id)
  * stack:break: stop or skip current lookup operation

These abstract instructions are included in the ouput of the generated
XSLT encoder, together with Iso-Java instructions. They get interpreted
in an intermediate step between the encoding from XML to Iso-B4J and the
conversion of Iso-B4J to binary. That transform, located at:

  src/powerpack/encoder/ReplaceStacksInIsoB4J.xsl

runs during the 'encoding' task of the Ant build file and converts these
abstract instructions into Iso-B4J. In the resulting binary files, these
stack actions are replaced with short integer values which allow the Java
decoder to create and manage the stacks efficiently using fixed-size arrays.