Add base XSLT Transform for generation of decoders in the powerpack
This is the extension of the base XSLT Transform with common processing
for the generation of the Java decoder and test decoder. It handles
aspect specific to the powerpack extension: the generation of code
for the stacks used in the resolution of Id/IdRef matching as well
as the decoding of inline files into an array of bytes.