Add XSLT Transform which generates the test Java decoder in the powerpack
This transform completes the resolution of Id/IdRef matching and file
inlining done in the base transform XsdToJavaDecoderBase.xsl, for the
implementation of the test Java decoder. Instead of running semantic
actions, the test decoder outputs XML with decoded values. In the case
of the inlined file, only the number of bytes read is printed, not
the whole contents of the file.