Add XML Schema for the basic subset of XML Schema used in B4J tools
This XML Schema defines a grammar for the basic subset of XML Schema
used in B4J tools to generate encoders and decoders.

While the comment at the top of the file claims that the schema
also defines the grammar for annotations with semantic actions,
only the xsd:annotation with its children xsd:documentation and
xsd:appinfo elements are defined here.

The extensions for B4J config and annotations are defined in two
separate schemas found in the same folder:

  * B4JConfig.xsd for the config:root element
  * B4JActionsForJava.xsd for the a4j:* elements for semantic actions

These definitions are combined in a third schema, also found in the
same folder: src/schema/BasicXmlSchemaPlusSemanticActions.xsd, which
includes BasicXmlSchema.xsd to extend it with elements from the two
namespaces for config:root and a4j:* elements, from imported schemas.

Comments about definitions which had to be done in the including schema
are referring to the file BasicXmlSchema.xsd.