Define color attribute as an enumeration
Previously, the byte value had to be provided directly in the XML document.
Now, the string values defined in the enumeration will make the color
attribute more explicit in the source XML document. This does not change
anything on the side of the decoder, which will still receive a byte value,
not the string value.

Only three color names were included here, which is sufficient for testing
purpose. A whole list of color names can be included in the same way.