Add base XSLT Transform which browses custom XML Schema files
The same base transform is used to browse XML Schema provided as input
for the generation of encoders and decoders. This process is described
on slide 12 of the presentation doc/overview/overview.pdf.

This is the whole implementation of the transform, not just a header.
It handles the parsing of the XML Schema structure, including the config
annotation, but not annotations for semantic actions, which are only used
for the generation of Java decoders, not for encoders. Specific steps are
implemented as empty templates at the end of this file, to be overriden
in more specific templates for the generation of encoders and decoders.