Add implementation of XSLT Transform which generates encoders
This transform implements the specific steps to generate the XSLT encoder.
It relies on the base transform XsdBrowsingTemplateMethod.xsl, located at
src/browsing/XsdBrowsingTemplateMethod.xsl, for the browsing of the XML
Schema provided as input. This import is located in the header part of
the transform, in the file src/encoder/XsdToXsltEncoder.xsl.