Add implementation of base XSLT Transform for generation of Java decoders
This is the base implementation for the generation of both the Java
decoders and test decoders. It relies itself on a base transform
for the browsing of the XML Schema provided as input, located at
src/browsing/XsdBrowsingTemplateMethod.xsl. The corresponding import
is done in the header of the transform, XsdToJavaDecoderBase.xsl found
the same folder.

The generated Java decoder is generated in an XML format with the
same structure as the Java instructions at this step, which gets
converted to actual Java in the next step of the processing, in
the transform IsoJavaToJava.xsl found in the same folder.