Add dependency: fplib version 1.6
fplib is a Fixed Point Math library for the IAppli (CLDC) Java platform.

  > The package works on fixed-point 32 bit ints,
  > with 16 bits of integer and 16 bits of fraction.
  > — http://www.beartronics.com/imode/fplib/

Author:

  Henry Minsky (hqm@ai.mit.edu)
  [1] http://www.beartronics.com
  [2] http://sourceforge.net/users/hqm/
  [3] https://github.com/hminsky

Project Home Page:

  [4] http://www.beartronics.com/imode/fplib/
  [5] http://sourceforge.net/projects/bearlib/

Source Code:

  [6] http://www.beartronics.com/imode/fplib/FP.java
  [7] http://www.beartronics.com/imode/fplib/FP.java.html
  [8] http://sourceforge.net/project/showfiles.php?group_id=24849&package_id=20439

API Documentation:

  [9] http://www.beartronics.com/imode/fplib/javadoc/

License:

  Artistic License
  [10] http://www.opensource.org/licenses/artistic-license.html

Some of these pages are no longer online, but have been preserved
by the Web Archive:

  * Author:
  [2] https://web.archive.org/web/20041101122237/http://sourceforge.net/users/hqm/

  * Projet Home Page:
  [4] https://web.archive.org/web/20080225155718/http://www.beartronics.com/imode/fplib/
  [5] https://web.archive.org/web/20041223182856/http://sourceforge.net/projects/bearlib/

  * Source Code:
  [6] https://web.archive.org/web/20070902073202/http://www.beartronics.com/imode/fplib/FP.java
  [7] https://web.archive.org/web/20080224010512/http://www.beartronics.com/imode/fplib/FP.java.html
  [8] https://web.archive.org/web/20080229003857/http://sourceforge.net/project/showfiles.php?group_id=24849&package_id=20439

  API Documentation:
  [9] https://web.archive.org/web/20080225155705/http://www.beartronics.com/imode/fplib/javadoc/

Version 1.6 of the library is not available from any of the two home pages:

  - the latest version on beartronics.com was v1.1.1.1 (2004-06-25)
  - the latest version on sourceforge.com was v1.0 (2001-06-03)

I have found it shipped as part of another project by Paulo Vieira
and Henry Minsky, still hosted on Sourceforge:

  Proteus J2ME Browser
  https://sourceforge.net/projects/protheus/

It is found inside the fplib.tgz archive from 2001-06-03 which was
also hosted on Sourceforge at the time:

  + fplib.tgz
  |-- FPMath.java: unversioned, undated, similar class
  |-+ lib
    |-+ com
      |-+ beartronics
        |-- FP.java: @version $Id: FP.java,v 1.6 2001/04/05 07:40:17 hqm Exp $

The version 1.6 (2001-04-05) of FP.java is actually found in the archive
labelled 1.0 in the project, which is much older than the source code shown
on beartronics.com for version 1.1.1.1 (2004-06-25), available at the time.

Another similar class is found in the same package from 2001-06-03,
FPMath.java, which seems to be an even older, undated version of the code,
in a class called FPMath, without package.