Add project report prepared for my Master presentation (December 2004)
This is the final report for my final presentation for graduation
in the Master in Video Games and Interactive Media, following the
end of my joint internship at Pastagames and ExpWay.

This PDF document has been exported from the source document with
the .sxw extension, found in the same folder.

It provides a lot of context about the projects that motivated the
development of the tool, and the directions that were envisioned for
its future development.