Add source of slides for my final Master presentation
This is the source document for the PDF of the same name found in
the same folder. It was created in an early version of OpenOffice.

This presentation was made on December 22nd, 2004, for the final
presentation before graduation in the Master in Video Games and
Interactive Media from CNAM-ENJMIN in Paris and Angoulême.

A more detailed overview of the project can be found in doc/overview/,
while details of the context of the project can be found in the
written report in the same folder:

  EricBrechemier_Automatiser_le_passage_des_donnees_aux_objets_java.pdf