Add XSLT Transform which generates Ant build.xml for powerpack and core
Depending on whether the B4J is the powerpack extension or the core
of the tool, this transform includes or omits parts of the tasks
in the generated build.xml.

There is a separate Ant file to run this transform and generate
the main build.xml, ant-config.xml, also found in the root folder.
The default is to generate the full build.xml for the powerpack.
Generating the build.xml for the core requires to set the parameter
b4jVersion to a different value from the command line:

  ant -f ant-config.xml -Db4jVersion=standard