Replace XML Schema for Fixed Point Math alternative to float
This schema defines a single simple type, fp:float, as an extension
of xsd:float. In the B4J toolchain, it gets associated with a codec
to encode floating point values as Fixed Point integers.

The namespace used to refer to MathFP, the Java library used previously
for Fixed Point math in this project. The new namespace is more neutral,
which reflects the use of an alternate library for Fixed Point math, fplib.