Replace namespace used for fixed point codec with more generic one
This new version does not refer explicitly to MathFP anymore,
since we have a choice of at least two different libraries
for this codec, and MathFP will get replaced with fplib.

Note that while the namespace declaration has changed, it is still
not used in this schema, which has no import yet and continues to
use the xsd:float type and not the fp:float type at this point.