Use official Apache Ant namespace rather than no namespace
XML elements in an Ant build file can either be declared without
a namespace, which is treated like the special namespace '#default'
in XSLT, or the XML elements can be placed in the official namespace
'antlib:org.apache.tools.ant'.

While the initial version of the build.xml (not saved in this
repository) did not use a namespace, the current version now
uses the official namespace, and I have updated this transform
accordingly. It takes the build.xml as input to read the user
properties for the target prefix and namespace for the empty
schema to be generated.