Prepare to support float and double values without Fixed Point integers
This is marked as a work in progress, with a comment which deprecates
the previous usage of the functions writeFloat() and writeDouble()
when expecting their encoding as Fixed Point integer values.

In the next version of this encoder class, the function writeFloatString()
will keep encoding Fixed Point integers, but writeFloat() and writeDouble()
will encode native Java floating point and double values instead.

In this work in progress version of writeFloat(), both versions are
actually done, because the new encoding code is not commented out and
runs before the old encoding code. This line should be commented out
until the next version, like in writeDouble():

  //_singleton.output.writeFloat(data);