Replace MathFP with Beartronics fplib for Fixed Point math values
I had to add the function stringToFP() to handle the conversion
from a string value to a fixed point value. This function is
present in MathFP, as int toFP(String s) but lacking in fplib,
which is initialized from a regular integer instead.

The current implementation of stringToFP() does not handle
a negative floating point value correctly: the sign is only
applied to the integer part, while the fractional part is
always parsed as a positive value, which results in a
discrepancy in the sum to form the full Fixed Point value.
Thus:

  int fpValue = fpIPart + fpFPart;
  return fpValue;

should be rewritten:

  int fpValue = fpIPart < 0 ? fpIPart - fpFPart: fpIPart + fpFPart;

Also, multiplication by -1 can be done directly on a Fixed Point
integer, which means that:

  FP.Mul(, -FP.ONE )
  FP.Mul(fpSinAngle, -FP.ONE )

can be simply written as:

  -fpCosAngle
  -fpSinAngle