Add URI resolver to locate XSLT transforms inside a JAR archive
I developed this class in support for the standalone packaging
of the encoder as a JAR, with classes and XSLT transforms within.
It is used in the main class of the standalone encoder, the class
PowerPackB4JEncoderWrapper.java found in the same folder.

The URI resolver is needed to translate the relative URIs of XSLT
transforms to include or import to the corresponding location in
the JAR, based on the location of the parent XSLT transform.