Add main class for standalone Java encoder
The standalone Java encoder is packaged by the new task
'packageEncoderLib' of the build.xml file, which creates
a JAR archive embedding Java code together with the XSLT
encoder generated for the user schema.

The main() method expects two parameters, the path to the
XML input file and the path to the B4J output file produced
by the encoding.

The class runs the same three transforms as the 'encoding' task
of the build.xml, which are chained to run as a single pipe:

  1. convert user XML data into Iso-B4J XML format
     and abstract stacks instructions for Id/IdRefs matching
  2. replace abstract stack instructions with Iso-B4J XML
  3. encode Iso-B4J XML to Binary for Java format

The resolution of relative URLs for XSLT transforms imported
in support of the requested transforms is done through the
custom URI resolver class JarURIResolver.java, found in the
same folder.

The encoding is triggered by the call to b4jEncode() from the main()
function, which creates an intermediate temporary file in the location
output/test/temp.b4j. This temporary file is then copied to the specified
output location through a loop which runs in convertInputToOutput().

This copy feels completely redundant. The chain of XSLT transforms should
write directly to the specified output location.

The intent was to allow the use of this encoder either as a standalone
executable from the command line or from a Java program, as a library.
Instead of providing the paths to input and output files, the library
would receive a DataInputStream and return a DataOutputStream. The
intermediate TEMP_B4J_FILE was a step in that direction, allowing
a Java program to get an InputStream out of b4jEncode(), instead of
having to specify an intermediate location on disk.