Add Fabien Delpiano as contributor for the concept
I met Fabien when he came to teach how to develop mobile games
in Java in the Master in Video Games and Interactive Media that
I attended as a student. He mentioned this recurrent need he had
to develop custom encoders and decoders to fit data and the parsing
code into the small memory of the target devices. I had some ideas
that I shared with him on the spot, using XSLT as my tool of choice.
We started a discussion around the concept of this tool, which lead
to my joint internship between Fabien's game studio, Pastagames,
and a partner company specializing in XML compression, ExpWay.