Simplify XSLT value templates, using {{ and }} to escape { and }
Previously, the { and } were left out of the XSLT value template,
resulting in some confusing sequences of punctuation. For example,
to obtain the following property in the generated Ant build.xml:

  <property name="jing.jar" value="${lib.path}/jing-20030619/jing.jar"/>

I previously wrote:

  <property name="jing.jar" value="{'${lib.path}/jing-20030619/jing.jar'}"/>

because the characters { ... } initialize a replacement expression called
Attribute Value Template in XSLT, and we just expect a literal { ... } here,
because the replacement is expected to be performed by Ant, not by XSLT.

This corresponds to the following pseudo-code:

  property/@name := "jing.jar";
  property/@value := xslt-attribute-value-template(
    '${lib.path}/jing-20030619/jing.jar'
  );

In other words, I am initializing an Attribute Value Template in which
I print a literal value, between the single quotes ' ... '.

To avoid the uncanny sequence {'${...'}, the XSLT recommendation allows
to treat {{...}} like a simple literal value { }. I can thus write simply:

  <property name="jing.jar" value="${{lib.path}}/jing-20030619/jing.jar"/>

Reference:

 “When an attribute value template is instantiated,
  a double left or right curly brace outside an expression
  will be replaced by a single curly brace. It is an error
  if a right curly brace occurs in an attribute value template
  outside an expression without being followed by a second right
  curly brace. A right curly brace inside a Literal in an
  expression is not recognized as terminating the expression.”

  — 7.6.2 Attribute Value Templates
  in XSL Transformations (XSLT) Version 1.0
  W3C Recommendation 16 November 1999
  https://www.w3.org/TR/1999/REC-xslt-19991116#attribute-value-templates